Reorganise highlight dir + some types
- Add common/ subdirectory to highlight/, for all utilities used between
  engines, special engines, models, and/or the engine manager
- Rename and/or split some files under highlight/common/
- Reorganise some highlight types
